Chapter 16 Finished
 Update pages with metadata for title

This commit updates the page components in the dashboard and login directories to include metadata for setting the title of each page. The metadata is defined using the `Metadata` interface from 'next'. The title is set based on the file name and some pages have a custom title. The layout component also includes metadata for the base URL and description.